add filters

+++ 08 PHP/homework/Space7Panda/5/app/src/delete.php
@@ -2,15 +2,17 @@
 
 require_once '../config/init.php';
 
-$item = $_GET['item'];
+if (isset($_GET['item'])) {
+    $item = $_GET['item'];
 
-$deleteQuerry = $db->prepare("
-DELETE FROM items 
-WHERE id = :item
-");
-
-$deleteQuerry->execute([
-'item' => $item,
-]);
+    $deleteQuerry = $db->prepare("
+        DELETE FROM items 
+        WHERE id = :item
+    ");
+    
+    $deleteQuerry->execute([
+        'item' => $item,
+    ]);
+}
 
 header('Location: ../index.php');
